8224210: Shenandoah: Refactor ShenandoahRootScanner to support scanning CSet codecache roots Reviewed-by: shade
